// src/app/models/order.ts
export class order {
  id: any;
  order_id: any | undefined;
  user_id: any | undefined;
  product_id: any | undefined;
  seller_id: any | undefined;
  quantity: any | undefined;
  color: any | undefined;
  product_price: any | undefined;
  total_product_price: any | undefined;
  total_price: any | undefined;
  paid: any | undefined;
  status: any | undefined;
  comment: any | undefined;
  user_required: any | undefined;
  created_at: any | undefined;
  updated_at: any | undefined;
  products: any;
}